requirements.txt
numpy
ipdb
scikit-learn
pytest
pytest-cov
pandas
matplotlib
torch
torchvision
tensorflow-gpu
tensorboardX
dataclasses
pyyaml
